+++ CHANGELOG.md
@@ -10,6 +10,7 @@ that you can set version constraints properly.
 
 #### [Unreleased][unreleased] -
 
+* Backfilled tests for Manage Intrinsic Size Relations methods
 * Backfilled test for Manage Relative Position family of methods
 * Changed Manage Relative Position family of methods to return collection of constraints
 * Backfilled test for `expand(fromMarginsOf:)` family of methods